[Test] 로그인 시 토큰 발급하는 테스트 분리
